Add log info to app title
Fixes #22
