+++ Fusion++.Engine/IO/TemporaryLogStore.cs
@@ -1,27 +1,19 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Security.AccessControl;
 using System.Security.Principal;
 
 namespace FusionPlusPlus.Engine.IO
 {
-	public class TemporaryLogStore : ILogStore
+	public class TemporaryLogStore : TransparentLogStore
 	{
-		private const string UTC_DATE_FORMAT = "yyyy.MM.dd.HH.mm.ss.ffff";
-
-		public TemporaryLogStore()
+		public TemporaryLogStore() : base("")
 		{
 			TopLevelPath = System.IO.Path.Combine(System.IO.Path.GetTempPath(), "Fusion++");
 			Path = System.IO.Path.Combine(TopLevelPath, DateTime.UtcNow.ToString(UTC_DATE_FORMAT));
 		}
 
-		public string TopLevelPath { get; }
-
-		public string Path { get; }
-
-		public void Prepare()
+		public override void Prepare()
 		{
 			if (!Directory.Exists(Path))
 			{
@@ -83,31 +75,6 @@ namespace FusionPlusPlus.Engine.IO
 			}
 		}
 
-		public string GetLogName(string path)
-		{
-			var result = System.IO.Path.GetFileName(path);
-			if (DateTime.TryParseExact(result, UTC_DATE_FORMAT, System.Globalization.CultureInfo.InvariantCulture, System.Globalization.DateTimeStyles.AssumeUniversal, out DateTime date))
-				result = date.ToShortDateString() + " " + date.ToLongTimeString();
-
-			var apps = new List<string>();
-			foreach (var subfolder in new string[] { "Default", "NativeImage" })
-			{
-				var folder = System.IO.Path.Combine(path, subfolder);
-				if (Directory.Exists(folder))
-					apps.AddRange(Directory.GetDirectories(folder).Select(System.IO.Path.GetFileName));
-			}
-
-			if (apps.Any())
-			{
-				apps = apps.Distinct().OrderBy(name => name).ToList();
-
-				if (apps.Count > 5)
-					apps = apps.Take(5).Union(new string[] { "..." }).ToList();
-
-				result += " " + string.Join(", ", apps);
-			}
-
-			return result;
-		}
+		public string TopLevelPath { get; }
 	}
 }
